requirements.txt
absl-py==0.8.1
annoy==1.16.3
appdirs==1.4.3
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.3
cx-Freeze==6.1
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
efficientnet==1.0.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
gast==0.2.2
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.25.0
gym==0.15.4
h5py==2.9.0
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==1.5.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.11.1
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-self-attention==0.42.0
keyboard==0.13.5
kiwisolver==1.1.0
labelImg==1.8.3
lightgbm==2.3.1
lmdb==0.98
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
MouseInfo==0.1.2
moviepy==1.0.1
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.16.3
oauthlib==3.1.0
objgraph==3.4.1
opencv-python==3.4.4.19
openpyxl==3.0.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.0.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.3
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.48
pydot==1.4.1
pygame==2.0.0.dev6
PyGetWindow==0.0.8
pyglet==1.3.2
Pygments==2.5.2
Pympler==0.8
PyMsgBox==1.0.7
pyparsing==2.4.5
pyperclip==1.7.0
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyRect==0.1.4
pyrsistent==0.15.7
PyScreeze==0.1.25
pytesseract==0.3.4
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.13.0
SQLAlchemy==1.3.17
stackimpact==1.2.7
tensorboard==1.14.0
tensorboard-plugin-wit==1.6.0.post2
tensorboardX==2.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
tesseract==0.1.3
testpath==0.4.4
torch==1.2.0+cu92
torchvision==0.4.0+cu92
tornado==6.0.3
tqdm==4.40.1
traitlets==4.3.3
urllib3==1.25.7
virtualenv==20.0.17
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
zipp==2.1.0
